+++ packages/sdk/test/ci/plt/module.test.ts
@@ -4,8 +4,11 @@ import {
     TokenAddAllowListOperation,
     TokenAddDenyListOperation,
     TokenBurnOperation,
+    TokenListUpdateEventDetails,
     TokenMintOperation,
     TokenOperationType,
+    TokenPauseEventDetails,
+    TokenPauseOperation,
     TokenRemoveAllowListOperation,
     TokenRemoveDenyListOperation,
     TokenTransferOperation,
@@ -22,23 +25,29 @@ import {
     serializeAccountTransactionPayload,
 } from '../../../src/pub/types.js';
 
-describe('PLT V1 parseModuleEvent', () => {
+describe('PLT parseModuleEvent', () => {
     const testEventParsing = (type: string, targetValue: number) => {
         it(`parses ${type} events correctly`, () => {
-            const target = TokenHolder.fromAccountAddress(
-                AccountAddress.fromBuffer(new Uint8Array(32).fill(targetValue))
-            );
+            const accountBytes = new Uint8Array(32).fill(targetValue);
+            const details: TokenListUpdateEventDetails = {
+                target: TokenHolder.fromAccountAddress(AccountAddress.fromBuffer(accountBytes)),
+            };
             const validEvent: EncodedTokenModuleEvent = {
                 tag: TransactionEventTag.TokenModuleEvent,
-                tokenId: TokenId.fromString('PLT'),
                 type,
-                details: Cbor.encode({ target }),
+                tokenId: TokenId.fromString('PLT'),
+                details: Cbor.encode(details),
             };
+            expect(validEvent.details.toString()).toEqual(
+                `a166746172676574d99d73a201d99d71a101190397035820${Buffer.from(accountBytes).toString('hex')}`
+            );
 
             const parsedEvent = parseModuleEvent(validEvent);
             expect(parsedEvent.type).toEqual(type);
-            expect(parsedEvent.details.target.type).toEqual('account');
-            expect(parsedEvent.details.target.address.decodedAddress).toEqual(new Uint8Array(32).fill(targetValue));
+            expect((parsedEvent.details as TokenListUpdateEventDetails).target.type).toEqual('account');
+            expect((parsedEvent.details as TokenListUpdateEventDetails).target.address.decodedAddress).toEqual(
+                new Uint8Array(32).fill(targetValue)
+            );
         });
     };
 
@@ -46,6 +55,21 @@ describe('PLT V1 parseModuleEvent', () => {
     testEventParsing('addDenyList', 0x16);
     testEventParsing('removeAllowList', 0x17);
     testEventParsing('removeDenyList', 0x18);
+
+    it('parses pause event', () => {
+        const details: TokenPauseEventDetails = { paused: true };
+        const validEvent: EncodedTokenModuleEvent = {
+            tag: TransactionEventTag.TokenModuleEvent,
+            tokenId: TokenId.fromString('PLT'),
+            type: 'pause',
+            details: Cbor.encode(details),
+        };
+        expect(validEvent.details.toString()).toEqual('a166706175736564f5');
+
+        const parsedEvent = parseModuleEvent(validEvent);
+        expect(parsedEvent.type).toEqual('pause');
+        expect(parsedEvent.details).toEqual({ paused: true });
+    });
 
     it('throws an error for invalid event type', () => {
         const invalidEvent: EncodedTokenModuleEvent = {
@@ -76,9 +100,19 @@ describe('PLT V1 parseModuleEvent', () => {
         };
         expect(() => parseModuleEvent(missingTargetEvent)).toThrowError(/{}/);
     });
+
+    it("throws an error if 'paused' is missing or invalid for pause events", () => {
+        const missingTargetEvent: EncodedTokenModuleEvent = {
+            tag: TransactionEventTag.TokenModuleEvent,
+            tokenId: TokenId.fromString('PLT'),
+            type: 'pause',
+            details: Cbor.encode({}),
+        };
+        expect(() => parseModuleEvent(missingTargetEvent)).toThrowError(/{}/);
+    });
 });
 
-describe('PLT v1 transactions', () => {
+describe('PLT transactions', () => {
     const token = TokenId.fromString('DKK');
     const testAccountAddress = TokenHolder.fromAccountAddress(AccountAddress.fromBuffer(new Uint8Array(32).fill(0x15)));
     // - d99d73: A tagged (40307) item with a map (a2) containing:
@@ -367,6 +401,37 @@ describe('PLT v1 transactions', () => {
               a1
                 6e72656d6f766544656e794c697374 a1
                   66746172676574 ${testAccountAddressCbor}
+            `.replace(/\s/g, ''),
+            'hex'
+        );
+
+        expect(payload.operations.toString()).toEqual(expectedOperations.toString('hex'));
+
+        const decoded = Cbor.decode(payload.operations);
+        expect(decoded).toEqual([removeDenyList]);
+
+        const ser = serializeAccountTransactionPayload({ payload, type: AccountTransactionType.TokenUpdate });
+        const serPayload = ser.slice(1);
+        const des = new TokenUpdateHandler().deserialize(Cursor.fromBuffer(serPayload));
+        expect(des).toEqual(payload);
+    });
+
+    it('(de)serializes pause operations correctly', () => {
+        const removeDenyList: TokenPauseOperation = {
+            [TokenOperationType.Pause]: true,
+        };
+
+        const payload = createTokenUpdatePayload(token, removeDenyList);
+
+        // This is a CBOR encoded byte sequence representing the pause operation:
+        // - 81: An array of 1 item
+        // - a1: A map with 1 key-value pair
+        //   - 65 7061757365 f5: Key "pause" (in UTF-8), value true
+        const expectedOperations = Buffer.from(
+            `
+            81
+              a1
+                65 7061757365f5
             `.replace(/\s/g, ''),
             'hex'
         );